fix: reenable accounts
